% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{get_.fds}
\alias{get_.fds}
\title{Get .fds}
\usage{
get_.fds()
}
\value{
If .fds can be found, it (a list of friends lists) will be returned
otherwise this produces an error.
}
\description{
Retrieves '.fds' object created by \code{\link{get_friends_recipe}()} and stored in
rtweet.recipes's special built-in environment.
}
\seealso{
Other .fds: \code{\link{get_friends_download}}
}
\concept{.fds}
