% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{get_.usr}
\alias{get_.usr}
\title{Get .usr}
\usage{
get_.usr()
}
\value{
If .usr can be found, it (a list of users data) will be returned
otherwise this produces an error.
}
\description{
Retrieves '.usr' object created by \code{\link{lookup_users_recipe}()} and stored in
rtweet.recipes's special built-in environment.
}
\seealso{
Other .usr: \code{\link{lookup_users_recipe}}
}
\concept{.usr}
