% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{get_followers_download}
\alias{get_followers_download}
\title{Get followers download}
\usage{
get_followers_download(x, ..., output = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Either a data frame or character vector containing user identifiers
for which follower IDs will be retrieved. See details for more information
about how this works.}

\item{...}{If \code{x} is a data frame this can be used to select columns
containing the appropriate user identifying information (user_id and/or
screen_name). This uses the tidyselect specification. If \code{x} is a
character vector, then the first unnamed or non-argument named value is
assumed to be labels (screen names) corresponding with \code{x}.}

\item{output}{Optionally supply a preexisting output vector (like that returned
by this function)–if NULL, the default, this function will start fresh.}

\item{verbose}{Whether the function should print information/status updates,
defaults to TRUE. Setting this to FALSE will silent most printing.}
}
\value{
Returns a list data frames with user and follower ID information. See
\code{\link[rtweet]{get_followers}} for more information.
}
\description{
Automate the collection of follower IDs for a large number of users (via
\code{\link[rtweet]{get_followers}})
}
\details{
This function attempts to retrieve up to 75,000 follower IDs every
15 minutes, sleeping between calls until Twitter's API rate
limit resets. If your API token is linked to your own Twitter APP and has
appropriate permissions to create a 'bearer token', then this function will
collect closer to 150,000 followers per 15 mintues.
}
\concept{followers}
