% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{get_friends_download}
\alias{get_friends_download}
\title{Get friends download}
\usage{
get_friends_download(x, new = FALSE)
}
\arguments{
\item{x}{A vector of user IDs or screen names for which friends lists will be
retrieved. See details for more information about how this works.}

\item{new}{Logical indicating whether to start a new data collection–if FALSE,
the default, then this function will pick up where any previous in-session
calls left off.}
}
\value{
Returns a list of friends lists; this list can be retrieved via
\code{\link{get_.fds()}} even if an error occurs (and the R session wasn't
restarted).
}
\description{
Automate friends-list collection for a large number of users (via \code{\link[rtweet]{get_friends()}})
}
\details{
This function attempts to retrieve friends lists for 15-30 users
every 15 minutes, sleeping between calls until Twitter's API rate limit
resets. If your API token is linked to your own Twitter APP and has
appropriate permissions to create a 'bearer token', then this function will
make 30 calls every 15 minutes. If the token cannot be used via bearer
authorization, then 15 calls are made every 15 minutes.
}
\seealso{
Other .fds: \code{\link{get_.fds}}
}
\concept{.fds}
