% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{lookup_users_download}
\alias{lookup_users_download}
\title{Lookup users download}
\usage{
lookup_users_download(x, .usr = NULL)
}
\arguments{
\item{x}{A vector of user IDs or screen names for which data will be looked up.
See details for more information about how this works.}

\item{.usr}{Optionally supply a preexisting output vector (like that returned
by this function)–if NULL, the default, this function will start fresh.}
}
\value{
Returns a list data frames of user data
}
\description{
Automate users data collection for a large number of users (via
\code{\link[rtweet]{lookup_users()}})
}
\details{
This function attempts to lookup data for 90,000 users every 15
minutes, sleeping between calls until Twitter's API rate limit resets.
}
\concept{.usr}
