% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet.recipes-package.R
\docType{package}
\name{rtweet.download-package}
\alias{rtweet.download}
\alias{rtweet.download-package}
\title{rtweet.download: Automate Large Downloads with 'rtweet'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Robust tools for automating large and/or time-consuming tasks
    involving the collection of Twitter data via 'rtweet' <https://rtweet.info>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/rtweet.download}
  \item Report bugs at \url{https://github.com/mkearney/rtweet.download/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (\href{https://orcid.org/0000-0002-0730-4694}{ORCID})

}
\keyword{internal}
