% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet.recipes-package.R
\docType{package}
\name{rtweet.recipes-package}
\alias{rtweet.recipes}
\alias{rtweet.recipes-package}
\title{rtweet.recipes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
