% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tweets.R
\name{search_tweets_download}
\alias{search_tweets_download}
\title{Search tweets download}
\usage{
search_tweets_download(q, n, ..., .stw = NULL)
}
\arguments{
\item{q}{Search query to be used to find matching tweets from the past 6-9
days. See \code{\link[rtweet]{search_tweets}}) for more information on
Twitter search query syntax.}

\item{n}{Number of desired tweets to return. See details for more information
about relevant rate limits.}

\item{...}{Other parameters are passed to
\code{\link[rtweet]{search_tweets}}).}

\item{.stw}{Optionally supply a preexisting output vector (like that returned
by this function)–if NULL, the default, this function will start fresh.}
}
\value{
Returns a list data frames of search data
}
\description{
Automate the data collection for large Twitter searches (via
\code{\link[rtweet]{search_tweets}})
}
\details{
This function attempts to search and collect data for up to 18,000
(when using the default rtweet authorization token) or 63,000 (when using
token generated from your own Twitter app with sufficient bearer token-
level permissions) statuses every 15 minutes, sleeping between calls unti
Twitter's API rate limit resets.
}
\concept{.stw}
