% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-token.R
\name{set_token}
\alias{set_token}
\title{Set token}
\usage{
set_token(x)
}
\arguments{
\item{x}{Either a token or path to a token. If path, the token is read using
\code{readRDS} (this is default {rtweet} behavior). If token, it is saved
in the current working directory as ".rtweet_token"}
}
\value{
The token is invisibly returned but more importantly the environment
variable "TWITTER_PAT" is set to point toward the saved token file. This
will be reset at the end of the session.
}
\description{
Stores Twitter API token information for the duration of the session
}
\examples{

## if your system already has an environment variable for an rtweet token,
## this will return the path
(pat <- Sys.getenv("TWITTER_PAT"))

## if your system doesn't have this environment variable OR if you wish to
## override this value, then enter the desired path or token object
#pat <- "/path/to/rtweet-token.rds"

## and then set the token for use for the remainder of the session
set_token(pat)

}
