% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{get_friends_download}
\alias{get_friends_download}
\title{Get friends download}
\usage{
get_friends_download(x, .fds = NULL, verbose = TRUE, sns = NULL)
}
\arguments{
\item{x}{A vector of user IDs or screen names for which friends lists will be
retrieved. See details for more information about how this works.}

\item{.fds}{Optionally supply a preexisting output vector (like that returned
by this function)–if NULL, the default, this function will start fresh.}

\item{verbose}{Whether the function should print information/status updates,
defaults to TRUE. Setting this to FALSE will silent most printing.}
}
\value{
Returns a list data frames with user and friend ID information
}
\description{
Automate friends-list collection for a large number of users (via
\code{\link[rtweet]{get_friends()}})
}
\details{
This function attempts to retrieve friends lists for 15-30 users
every 15 minutes, sleeping between calls until Twitter's API rate limit
resets. If your API token is linked to your own Twitter APP and has
appropriate permissions to create a 'bearer token', then this function will
make 30 calls every 15 minutes. If the token cannot be used via bearer
authorization, then 15 calls are made every 15 minutes.
}
\concept{.fds}
