% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet.recipes-package.R
\docType{package}
\name{rtweet.recipes-package}
\alias{rtweet.recipes}
\alias{rtweet.recipes-package}
\title{rtweet.recipes: Robust Automation Recipes for Collecting Twitter Data}
\description{
Tools and recipes for automating large scale collection of Twitter
    data via 'rtweet' <https://rtweet.info>
}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
